chore: extend incremental filename enforcement to team billing
